+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -78,5 +78,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-15T08:02:04.170Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-15T10:20:49.148Z")
     r"""The date and time the resource was last updated"""