+++ packages/orq-rc/RELEASES.md
@@ -3319,3 +3319,13 @@ Based on:
 - [python v3.14.0-rc.27] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.27] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.27 - packages/orq-rc
+
+## 2025-10-15 10:22:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.28] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.28] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.28 - packages/orq-rc